+++ src/nexus/NexusObject.cpp
@@ -25,14 +25,14 @@ int NexusObject::getObjectID()
 	return NexusObject::lastId++;
 }
 
-void NexusObject::addStretchConstraint(Particle* p1, Particle* p2, float distance, float stiffness) {
+void NexusObject::addStretchConstraint(Particle* p1, Particle* p2, double distance, double stiffness) {
 	constraints.push_back(mkU<StretchConstraint>(p1, p2, distance, stiffness));
 }
 
-void NexusObject::addBendingConstraint(Particle* p1, Particle* p2, Particle* p3, Particle* p4, float stiffness) {
+void NexusObject::addBendingConstraint(Particle* p1, Particle* p2, Particle* p3, Particle* p4, double stiffness) {
 	vec3 n1 = glm::normalize(glm::cross(p2->x - p1->x, p3->x - p1->x));
 	vec3 n2 = glm::normalize(glm::cross(p2->x - p1->x, p4->x - p1->x));
-	uPtr<BendingConstraint> d = mkU<BendingConstraint>(p1, p2, p3, p4, glm::acos(glm::clamp(glm::dot(n1, n2), -1.f, 1.f)), stiffness);
+	uPtr<BendingConstraint> d = mkU<BendingConstraint>(p1, p2, p3, p4, glm::acos(glm::clamp(glm::dot(n1, n2), -1.0, 1.0)), stiffness);
 	constraints.push_back(std::move(d));
 }
