+++ src/nexus/NexusObject.h
@@ -22,8 +22,8 @@ public:
 	~NexusObject();
 
 	void addParticle(uPtr<Particle> p);
-	void addStretchConstraint(Particle* p1, Particle* p2, float distance, float stiffness);
-	void addBendingConstraint(Particle* p1, Particle* p2, Particle* p3, Particle* p4, float stiffness);
+	void addStretchConstraint(Particle* p1, Particle* p2, double distance, double stiffness);
+	void addBendingConstraint(Particle* p1, Particle* p2, Particle* p3, Particle* p4, double stiffness);
 	void fixParticle(Particle* p);
 	const std::vector<uPtr<Particle>>& getParticles() const;
 	virtual void preComputeConstraints() = 0;
